ref: remove some Abstract components
